requirements/pure.txt
# Pure-python dependencies; pre-installed in vendor/

Django==1.3.1
python-memcached==1.48
django-tastypie==0.9.11
git+git://github.com/jeads/datasource@021797d4
